Make a task list request yo
